# api/openapi.yaml
openapi: 3.1.0
info:
  title: Push-Point API
  version: 0.1.0
  description: |
    Push-Point v2 HTTP API의 기계 원본(contract-first). 단일 사용자 앱이며 정적 Bearer API 키 1개(`PUSHPOINT_API_KEY`)로 인증한다.

    이 문서가 API의 단일 진실 원천이다 — `docs/v2/06-API-SPECIFICATION.md`(사람용 해설)와 다르면 이 파일이 우선한다.
    시간 필드는 전부 unix epoch 초 단위 정수다. `GET /debug/pprof/*`(Go 표준 프로파일링)는 서버가 서빙하지만 계약 대상이 아니다.
servers:
  - url: http://localhost:8420
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/links:
    get:
      operationId: listLinks
      tags: [links]
      summary: 링크 목록 조회
      description: >
        keyset 커서 페이지네이션 (OFFSET 금지). `(created_at, id)` 인덱스를 타므로
        링크 10만 건에서도 p99 < 50ms.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/TagFilter"
        - name: status
          in: query
          description: 링크 상태 필터
          schema:
            $ref: "#/components/schemas/LinkStatus"
      responses:
        "200":
          description: 링크 목록 페이지
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createLink
      tags: [links]
      summary: 링크 저장
      description: >
        `INSERT INTO links` + `INSERT INTO jobs(kind='scrape')`를 한 트랜잭션으로 커밋하고
        즉시 응답한다 (p99 < 50ms). 스크랩·태깅·썸네일은 백그라운드 잡으로 처리된다.
        `url_hash`(SHA-256(url)) 기반으로 멱등하다 — 클라이언트(Share Extension의 App Group
        로컬 큐)가 같은 요청을 재시도해도 중복 생성이 없다 (중복 시 200 `duplicate:true`).
        소프트 삭제된 링크의 URL을 다시 저장하면 같은 행을 복원(pending 복귀, note 교체,
        scrape 재-enqueue)하고 신규 저장처럼 201로 응답한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkInput"
      responses:
        "201":
          description: 신규 저장 완료 (scrape 잡 enqueue됨)
          content:
            application/json:
              schema:
                type: object
                required: [id, status, created_at]
                properties:
                  id:
                    type: integer
                  status:
                    $ref: "#/components/schemas/LinkStatus"
                  created_at:
                    $ref: "#/components/schemas/EpochSeconds"
              example:
                id: 1234
                status: pending
                created_at: 1752980000
        "200":
          description: >-
            동일 `url_hash`가 이미 존재 — 새로 만들지 않고 기존 id를 반환한다. 저장된 본문이
            서버 출처인데 요청이 클라이언트 본문(body_text)을 실어 오면 제목·설명·본문을 **1회
            보충**하고 태깅을 다시 돌린다(이미 클라이언트 본문이면 아무것도 하지 않는다).
            반복 호출은 같은 상태로 수렴하므로 재시도는 여전히 안전하다.
          content:
            application/json:
              schema:
                type: object
                required: [id, duplicate]
                properties:
                  id:
                    type: integer
                  duplicate:
                    type: boolean
                    const: true
              example:
                id: 987
                duplicate: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/links/{id}:
    parameters:
      - $ref: "#/components/parameters/LinkId"
    get:
      operationId: getLink
      tags: [links]
      summary: 링크 상세 조회
      description: 목록 항목 전체 필드 + author, published_at, duration_sec, word_count, lang, error, jobs 요약.
      responses:
        "200":
          description: 링크 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateLink
      tags: [links]
      summary: 링크 수정 (메모/태그)
      description: >
        `note` 전달 시 메모 교체. `tags`(태그 이름 배열) 전달 시 이 링크의 태그를 전체 교체한다 —
        추가분은 `link_tags(source='manual', confidence=NULL)` + `tag_feedback(added)`,
        제거분은 `tag_feedback(removed)` 기록. tag_feedback은 M5 재랭킹 학습 데이터다.
        `tags`가 `null`이거나 생략되면 태그를 유지하고, 빈 배열 `[]`이면 전체 제거로 처리한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkUpdateInput"
      responses:
        "200":
          description: 수정 반영된 링크 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteLink
      tags: [links]
      summary: 링크 삭제 (소프트)
      description: 소프트 삭제 — `deleted_at`만 기록하며 행은 남는다. 이후 목록/검색/상세에서 제외된다.
      responses:
        "204":
          description: 삭제 완료 (본문 없음)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/links/{id}/retry:
    post:
      operationId: retryLink
      tags: [links]
      summary: 실패 링크 재시도
      description: >
        `status='failed'`인 링크의 잡을 다시 enqueue한다. 링크는 `pending`으로 되돌아가고
        워커가 처음부터 다시 처리한다. `failed` 상태가 아닌 링크는 400.
      parameters:
        - $ref: "#/components/parameters/LinkId"
      responses:
        "202":
          description: 재-enqueue 수락
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id:
                    type: integer
                  status:
                    $ref: "#/components/schemas/LinkStatus"
              example:
                id: 1234
                status: pending
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tags:
    get:
      operationId: listTags
      tags: [tags]
      summary: 태그 사전 조회
      description: >-
        통제된 태그 사전 전체 + 사용 수(`link_count`, 삭제되지 않은 링크 기준).
        `facet`은 여기에만 실린다 — `LinkTag`에는 없으므로 클라이언트가 이 응답을 `Map<tagId, facet>`으로
        캐시해 목록 행의 칩 색을 해석한다 (링크당 태그 수만큼 페이로드가 늘어나는 것을 피한다).
      responses:
        "200":
          description: 태그 사전
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createTag
      tags: [tags]
      summary: 태그 생성
      description: >-
        `name`은 대소문자 무시 UNIQUE — 중복이면 400.
        `facet`은 optional이며 생략하면 `neutral`이다 (클라이언트가 칩 색을 고르는 유일한 입력).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/TagInput"
                - required: [name]
      responses:
        "201":
          description: 생성된 태그
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tags/{id}:
    parameters:
      - $ref: "#/components/parameters/TagId"
    patch:
      operationId: updateTag
      tags: [tags]
      summary: 태그 수정
      description: >
        `name`, `aliases`, `facet` 모두 optional (전달한 필드만 교체). `aliases`는 규칙 기반 태거의 매칭 대상 —
        alias를 잘 채우는 것이 태깅 정확도를 올리는 가장 싼 방법이다.
        `facet`은 태그 분류 축이며, 이 값을 바꾸면 모든 클라이언트에서 그 태그의 칩 색이 바뀐다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "200":
          description: 수정된 태그
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteTag
      tags: [tags]
      summary: 태그 삭제
      description: 사전에서 제거한다. `link_tags` 연결은 FK CASCADE로 함께 삭제된다.
      responses:
        "204":
          description: 삭제 완료 (본문 없음)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: 통계 조회
      description: iOS 위젯용 데이터를 한 번의 호출로 제공한다 (구현 완료 — 위젯 활용은 M6).
      responses:
        "200":
          description: 통계
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /thumbs/{dir}/{file}:
    get:
      operationId: getThumb
      tags: [thumbs]
      summary: 썸네일 정적 서빙
      description: >
        `data/thumbs/` 아래의 썸네일 JPEG(단일 사이즈, 최대 폭 640px, q80)를 그대로 서빙한다.
        경로 형식은 `{url_hash 앞 2자리}/{url_hash}.jpg` — 목록/상세 응답의 `thumb_url`이
        정확히 이 경로를 가리킨다. 인증 면제 — Tailscale이 네트워크 경계를 이루고,
        iOS AsyncImage가 커스텀 헤더를 지원하지 않기 때문이다.
      security: []
      parameters:
        - name: dir
          in: path
          required: true
          description: "url_hash 앞 2자리 (hex)"
          schema:
            type: string
            pattern: "^[0-9a-f]{2}$"
          example: a3
        - name: file
          in: path
          required: true
          description: "`{url_hash}.jpg` 파일명"
          schema:
            type: string
            pattern: "^[0-9a-f]{64}\\.jpg$"
          example: a3f1b2c4d5e6f7a8a3f1b2c4d5e6f7a8a3f1b2c4d5e6f7a8a3f1b2c4d5e6f7a8.jpg
      responses:
        "200":
          description: 썸네일 이미지
          content:
            image/jpeg:
              schema:
                type: string
                contentMediaType: image/jpeg
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "정적 API 키 1개 (`PUSHPOINT_API_KEY`). `Authorization: Bearer {key}`"
  parameters:
    LinkId:
      name: id
      in: path
      required: true
      description: 링크 id
      schema:
        type: integer
...
    Cursor:
      name: cursor
      in: query
      description: 이전 응답의 `next_cursor` 값. 첫 페이지는 생략. 불투명(opaque) 토큰.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: 페이지 크기
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    TagFilter:
      name: tag
      in: query
      description: 태그 이름 필터
      schema:
        type: string
  responses:
    BadRequest:
      description: 잘못된 요청 (`invalid_input`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: invalid_input
              message: url is required
    Unauthorized:
      description: API 키 누락 또는 불일치 (`unauthorized`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: unauthorized
              message: invalid api key
    NotFound:
      description: 리소스 없음 — 소프트 삭제된 링크 포함 (`not_found`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: link not found
    InternalError:
      description: >-
        서버 내부 오류 (`internal`) — 핸들러가 처리하지 못한 에러의 공통 종착점이다.
        `GET /healthz`만 이 응답이 없다 (조건 없는 단일 반환이라 실패 경로가 없다).
        `GET /thumbs/{dir}/{file}`은 인증만 면제일 뿐 500 면제는 아니다 — 파일 열기·stat 실패 시 500을 낸다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: internal
              message: internal server error
  schemas:
    EpochSeconds:
      type: integer
      description: unix epoch seconds
...
    LinkStatus:
      type: string
      description: 링크 처리 파이프라인 상태
      enum: [pending, scraping, tagging, done, failed]
...
    LinkDetail:
      description: 링크 상세 — 목록 항목 전체 필드 + 메타·에러·잡 요약
      allOf:
        - $ref: "#/components/schemas/Link"
        - type: object
          required: [author, published_at, duration_sec, word_count, lang, summary, error, jobs]
          properties:
            author:
              type: string
            summary:
              type: string
              description: >-
                본문에서 고른 핵심 문장 2~3개를 개행으로 이은 추출식 요약 (LLM 없이 원문
                문장을 그대로 선택하므로 환각 0). 본문이 얇거나 산문이 부족하거나 description과
                사실상 같으면 빈 문자열이며, 그때 클라이언트는 요약 영역을 그리지 않는다.
                목록(Link)·검색(SearchResult)에는 싣지 않는다 — 요약은 원문 대체재가 아니라
                "열까 말까"의 판단 보조다
            published_at:
              $ref: "#/components/schemas/EpochSecondsNullable"
            duration_sec:
              type: [integer, "null"]
            word_count:
              type: [integer, "null"]
            lang:
              type: string
            error:
              type: string
              description: 최종 실패 사유. 실패 없으면 빈 문자열
            jobs:
              type: object
              description: >-
                이 링크에 연결된 잡 상태 요약. 해당 kind의 잡이 아직 없으면 필드 생략
                (scrape 잡은 저장 트랜잭션에서 항상 생성되므로 필수). thumb은 best-effort —
                failed여도 링크 status에 영향 없음
              required: [scrape]
              properties:
                scrape:
                  $ref: "#/components/schemas/JobStatus"
                tag:
                  $ref: "#/components/schemas/JobStatus"
                thumb:
                  $ref: "#/components/schemas/JobStatus"
    LinkInput:
      type: object
      description: >-
        링크 저장 요청. title·description·body_text는 **클라이언트 캡처**용 optional 필드다 —
        서버가 fetch할 수 없는 페이지(SPA·봇 차단·로그인 벽)에서 이미 렌더된 콘텐츠를 가진
        클라이언트(브라우저 확장·iOS Share Extension)가 함께 보낸다. 서버는 길이를 **절단**할 뿐
        초과를 400으로 거부하지 않는다(클라이언트가 룬/바이트 경계를 서버와 맞출 방법이 없다).
      required: [url]
      properties:
        url:
          type: string
        note:
          type: string
          description: 개인 메모 (optional)
        title:
          type: string
          description: 클라이언트가 캡처한 제목 (optional, 512바이트로 절단)
        description:
          type: string
          description: 클라이언트가 캡처한 설명 (optional, 2048바이트로 절단)
        body_text:
          type: string
          description: >-
            클라이언트가 추출한 본문 평문 (optional, 32KB로 절단). 값이 있으면 서버는 이 링크의
            본문 출처를 'client'로 표시하고, 이후 스크랩이 제목·설명·본문을 덮어쓰지 않는다.
            태깅·요약이 이 본문을 입력으로 쓴다 — 그래서 스크랩이 실패해도 태그·요약이 나온다.
    LinkUpdateInput:
      type: object
      description: 링크 수정 요청 — note는 교체, tags는 전체 교체 (부분 추가/삭제 API 없음)
      properties:
        note:
          type: string
        tags:
          type: array
          description: 태그 이름 배열. 존재하지 않는 태그 이름이면 400
          items:
            type: string
    LinkPage:
      type: object
      description: 커서 페이지네이션 응답 래퍼
      required: [links, next_cursor]
      properties:
        links:
          type: array
          items:
            $ref: "#/components/schemas/Link"
        next_cursor:
          type: [string, "null"]
          description: 다음 페이지 커서. null이면 마지막 페이지
...
    Tag:
      type: object
      description: 태그 사전 항목 + 사용 수
      required: [id, name, aliases, link_count, facet]
      properties:
        id:
          type: integer
        name:
          type: string
          description: 태그 이름 (대소문자 무시 UNIQUE)
        aliases:
          type: array
          description: 동의어·영문/한글 표기 — Phase A 문자열 매칭의 재료
          items:
            type: string
        link_count:
          type: integer
          description: 이 태그가 붙은 (삭제되지 않은) 링크 수
        facet:
          $ref: "#/components/schemas/TagFacet"
...